разделил старый модуль поиска на слушатель ввода и модуль поиска
