+++ README.md
@@ -80,6 +80,8 @@ init project
 1. написал скрипт переключения активной страницы пагинатора, пока без перезагрузки выдачи. Попутно нашел ошибку - список с со страницами пагинатора отрисовывался не в предназначенном для него контейнере, а рядом с этим контейнером
 2. добавил условие, чтобы активная страница переключалась только при нажатии на неактивную страницу. Чтобы лишний раз не перезагружать выдачу при мисклике
 3. при переключение активных страниц, последнее значение записвается в localStorage. При перезагрузке страницы выводятся результаты поиска с учетом последних значений фильтров
+4. разделил старый модуль поиска на слушатель ввода в поиск и новый модуль поиска. Теперь можно использовать модуль поиска в модуле изменения активной страницы пагинатора
+
