+++ src/js/searchInput.js
@@ -1,28 +1,16 @@
 import listRepModule from '@js/listRepLoader';
 import makeRepsListModule from '@js/makeRepsList';
 import makePaginatorModule from '@js/paginatorLoader';
-let searchInput = document.querySelector('#searchInput');
 
-let searchModule = function() {    
-    //создаем событие для ввода запроса
-    searchInput.addEventListener ('keydown', function (evt) {
-        if (evt.keyCode === 13) {
-            if (searchInput.value === '') {
-                searchInput.placeholder = 'репозиторий не указан :(';
-            } else {                
-                let requestURL = 'https://api.github.com/search/repositories?q=' + searchInput.value + '&sort=stars&page=1&per_page=1';
-                //заменить количество репов на 10 ^
-                localStorage.setItem('searchName', searchInput.value.toString());
-                localStorage.setItem('page', '1');
-                makeRepsListModule(requestURL, 2000)
-                .then(dataList => {
-                    listRepModule(dataList.listItems);
-                    console.log(dataList.totalCount);
-                    makePaginatorModule(dataList.totalCount);
-                }).catch(err => console.log(err));
-            };
-        };
-    });
+let searchModule = function(searchName, searchPage) { 
+    let requestURL = 'https://api.github.com/search/repositories?q=' + searchName + '&sort=stars&order=desc&page=' + searchPage + '&per_page=2';   
+    //заменить количество репов на 10 ^
+    makeRepsListModule(requestURL, 2000)
+    .then(dataList => {
+        listRepModule(dataList.listItems);
+        console.log(dataList.totalCount);
+        makePaginatorModule(dataList.totalCount);
+    }).catch(err => console.log(err));
 };
 
 export default searchModule;
